pages/requirements.txt
folium==0.15.1
haversine==2.8.1
inflection==0.5.1
numpy==1.24.3
pandas==2.0.3
Pillow==10.2.0
plotly==5.9.0
streamlit==1.31.1
streamlit_folium==0.18.0
